requirements.txt

numpy
networkx
pyrr
pyyaml
pillow
scipy
tqdm
trimesh==3.21.7
matplotlib
simple-3dviz==0.7.0
num2words
nltk
transformers==4.25.1
clip==0.2.0
seaborn
wandb
hydra-core
tensorboard
omegaconf
loguru
open3d==0.17.0
opencv-python==4.8.0.74
tabulate==0.9.0
torchtext==0.15.2
num2words==0.5.12
einops
clip
smplx
